test(server): separate the socket claims that are POSIX from the ones that are not
Each of these was judged on whether it pins a property of POSIX or a claim
about this server, because the answer differed per test and only one of them
turned out to be a product bug (fixed separately, in clearStaleSocket).

POSIX properties, now behind the same !windows constraint socketmode_unix_test.go
already uses. The socket's permission mode is not a claim this server makes on
Windows: there are no POSIX bits on a socket there, access follows the
directory ACL, and bindUnixSocket already warns rather than pretending
otherwise, so asserting 0660 was demanding the platform be something it is not.
That moves the mode half of the serving test and the default-mode test.
TestClearStaleSocket_AnUnreadablePath moves for a subtler reason: refusing an
unanswered lstat is a general property, but arranging for one is not. It puts
the socket under a regular file, which POSIX answers with ENOTDIR and Windows
answers with ERROR_PATH_NOT_FOUND, which Errno.Is maps to fs.ErrNotExist. There
the question is answered rather than unanswered, and nil is the correct reply.

What is left in listen_test.go runs everywhere, including binding a unix socket
and holding a real HTTP conversation over it, which Windows does support.

Two cases inside TestListenUnix_RefusesAPathItCannotOwn get a marked skip
instead of a move, because those four are one question and reading them apart
is worse than reading them with a reason attached. The directory-is-a-file case
is the ENOTDIR problem again. The path-length case is ours only because the
POSIX path binds under a staging name and publishes with link(2), which has no
address limit; Windows binds the operator's path directly and the kernel
refuses an over-long one itself. Nothing binds that should not, so this is a
message-quality gap on Windows rather than a defect: the operator gets "bind:
invalid argument" instead of our sentence naming the limit.
